💫 [Feat] 가게의 미션을 도전 중인 미션에 추가 API 구현 #9

+++ spring/src/main/java/umc/study/repository/MissionRepository/MissionRepository.java
@@ -3,5 +3,5 @@ package umc.study.repository.MissionRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import umc.study.domain.Mission;
 
-public interface MissionRepository extends JpaRepository<Mission, Integer>, MissionRepositoryCustom {
+public interface MissionRepository extends JpaRepository<Mission, Long>, MissionRepositoryCustom {
 }
